+++ gulpfile.js
@@ -1,38 +1,61 @@
 const gulp = require('gulp');
 const args = require('yargs').argv;
-const webpackStream = require('webpack-stream');
-const webpack = require('webpack');
 const path = require('path');
+const hash = require('gulp-hash-filename');
 // All gulp plugins are automatically loaded into this constant, provided they're in package.json
 const plugins = require('gulp-load-plugins')({
     rename: {
         'gulp-clean-css': 'cleanCss'
     }
 });
+const webpackStream = require('webpack-stream');
+const webpack = require('webpack');
+const HtmlReplaceWebpackPlugin = require('html-replace-webpack-plugin');
 
 const packageJson = require('./package.json');
+
+const hashManifest = {};
+
+gulp.task('clean-css',  () => {
+    return gulp.src('dist/css')
+        .pipe(plugins.clean());
+});
 
 /**
  * Process the SASS files
  */
-gulp.task('sass', () => {
+gulp.task('sass', ['clean-css'], () => {
     return gulp.src(['src/scss/main.scss'])
         .pipe(plugins.plumber())
         .pipe(plugins.if(args.debug, plugins.debug({title: 'SASS'})))
         .pipe(plugins.sourcemaps.init())
         .pipe(plugins.sass().on('error', plugins.sass.logError))
         .pipe(plugins.autoprefixer())
-        // .pipe(plugins.if(!args.dev, plugins.cleanCss()))
-        .pipe(plugins.concat('styles.css'))
-        .pipe(plugins.cssimport())
+        .pipe(plugins.if(!args.dev, plugins.cleanCss()))
+        .pipe(hash({
+            "format": "styles.{hash}{ext}"
+        }))
+        .pipe(plugins.intercept((file) => {
+            const fileParts = file.path.split('\\');
+            const filename = fileParts[fileParts.length - 1]
+            const hash = filename.split('.')[1];
+            hashManifest.styles = hash;
+            return file;
+        }))
+        // .pipe(plugins.concat('styles.' + hashManifest.styles + '.css'))
         .pipe(plugins.sourcemaps.write('.'))
         .pipe(gulp.dest('dist/css'));
+});
+
+gulp.task('clean-js',  () => {
+    return gulp.src('dist/js')
+        .pipe(plugins.clean());
 });
 
 /**
  * Process the JS files
  */
-gulp.task('webpack', () => {
+gulp.task('js', ['clean-js'], () => {
     const webpackPlugins = [];
 
     if (!args.dev) {
@@ -43,10 +66,25 @@ gulp.task('webpack', () => {
         }));
     }
 
+    webpackPlugins.push(function () {
+        this.plugin('done', stats => {
+            const assetsByChunkName = stats.toJson().assetsByChunkName.main;
+            const scriptParts = assetsByChunkName[0].split('.')
+            hashManifest.scripts = scriptParts[1];
+        });
+    });
+
+    webpackPlugins.push(new HtmlReplaceWebpackPlugin([
+        {
+            pattern: '[[applicationVersion]]',
+            replacement: packageJson.version
+        }
+    ]));
+
     //App scripts
     return gulp.src('src/js/*.js')
         .pipe(plugins.plumber())
-        .pipe(plugins.if(args.debug, plugins.debug({title: 'WEBPACK'})))
+        .pipe(plugins.if(args.debug, plugins.debug({title: 'JS'})))
         .pipe(webpackStream({
             resolve: {
                 modules: [
@@ -62,21 +100,26 @@ gulp.task('webpack', () => {
                 }]
             },
             output: {
-                filename: 'scripts.js'
+                filename: 'scripts.[hash].js'
             }
         }, webpack))
+        .pipe(plugins.replace('[[applicationVersion]]', () => {
+            return packageJson.version;
+        }))
         .pipe(gulp.dest('dist/js'));
 });
 
-gulp.task('js', ['webpack'], () => {
-    return gulp.src('dist/js/scripts.js')
-        .pipe(plugins.plumber())
-        .pipe(plugins.if(args.debug, plugins.debug({title: 'JS'})))
-        .pipe(gulp.dest('dist/js'));
-});
-
-gulp.task('html', () => {
-    return gulp.src('src/*.html')
+gulp.task('html', ['js', 'sass'], () => {
+    return gulp.src('src/index.html')
+        .pipe(plugins.replace('[[scriptsHash]]', () => {
+            return hashManifest.scripts;
+        }))
+        .pipe(plugins.replace('[[stylesHash]]', () => {
+            return hashManifest.styles;
+        }))
+        .pipe(plugins.replace('[[applicationVersion]]', () => {
+            return packageJson.version;
+        }))
         .pipe(gulp.dest('dist'))
 });
 
@@ -100,11 +143,22 @@ gulp.task('vendorFiles', () => {
     .pipe(gulp.dest('dist/vendor'))
 });
 
+gulp.task('generate-service-worker', ['js', 'sass'], () => {
+    return gulp.src(['src/sw.js'])
+        .pipe(plugins.replace('[scriptsHash]', () => {
+            return hashManifest.scripts;
+        }))
+        .pipe(plugins.replace('[stylesHash]', () => {
+            return hashManifest.styles;
+        }))
+        .pipe(gulp.dest('dist'))
+});
+
 /**
  * Build the application.
  * This will be run when developing
  */
-gulp.task('build', ['sass', 'js', 'html', 'images', 'other', 'vendorFiles']);
+gulp.task('build', ['sass', 'js', 'html', 'images', 'other', 'vendorFiles', 'generate-service-worker']);
 
 /**
  * Watch for file changes and trigger the build task, the update the web page
@@ -120,5 +174,9 @@ gulp.task('default', ['build'], function () {
         gulp.watch([
             "src/*.html"
         ], ['html']);
+
+        gulp.watch([
+            "src/sw.js"
+        ], ['generate-service-worker']);
     }
 });